add refactoring Profile epic
