+++ FakeNFT/Models/Network/Epic Profile/InternetConnectionManager.swift
@@ -1,0 +1,34 @@
+//
+//  InternetConnectionManager.swift
+//  FakeNFT
+//
+//  Created by Ann Goncharova on 25.11.2023.
+//
+
+import UIKit
+import SystemConfiguration
+// базовая проверка наличия интернет-подключения на девайсе (можно ли установить соединение с сетью и является ли оно доступным)
+public class InternetConnectionManager {
+
+    private init() { }
+
+    public static func isConnectedToNetwork() -> Bool {
+        var zeroAddress = sockaddr_in()
+        zeroAddress.sin_len = UInt8(MemoryLayout.size(ofValue: zeroAddress))
+        zeroAddress.sin_family = sa_family_t(AF_INET)
+        guard let defaultRouteReachability = withUnsafePointer(to: &zeroAddress, {
+            $0.withMemoryRebound(to: sockaddr.self, capacity: 1) {
+                SCNetworkReachabilityCreateWithAddress(nil, $0)
+            }
+        }) else {
+            return false
+        }
+        var flags = SCNetworkReachabilityFlags()
+        if !SCNetworkReachabilityGetFlags(defaultRouteReachability, &flags) {
+            return false
+        }
+        let isReachable = (flags.rawValue & UInt32(kSCNetworkFlagsReachable)) != 0
+        let needsConnection = (flags.rawValue & UInt32(kSCNetworkFlagsConnectionRequired)) != 0
+        return (isReachable && !needsConnection)
+    }
+}
