+++ FakeNFT/Models/Network/Epic Profile/ProfileModel.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-struct ProfileNetworkModel: Codable {
+struct ProfileModel: Codable {
     let name: String
     let avatar: String
     let description: String
