+++ FakeNFT/Models/Network/Epic Profile/ProfileUpdateDelegate.swift
@@ -1,0 +1,12 @@
+//
+//  ProfileUpdateDelegate.swift
+//  FakeNFT
+//
+//  Created by Ann Goncharova on 25.11.2023.
+//
+
+import Foundation
+
+protocol ProfileUpdateDelegate: AnyObject {
+    func update()
+}
