+++ FakeNFT/Scenes/Profile/ProfileFavorites/FavoritesViewController.swift
@@ -51,6 +51,7 @@ final class FavoritesViewController: UIViewController, UIGestureRecognizerDelega
     }
     
     override func viewDidAppear(_ animated: Bool) {
+        super.viewDidAppear(animated)
         if badConnection { viewModel.getLikedNFTs(likedIDs: likedIDs) }
     }
