+++ FakeNFT/Scenes/Profile/ProfileMyNFT/MyNFTViewModel.swift
@@ -16,6 +16,7 @@ protocol MyNFTViewModelProtocol: AnyObject {
     var authors: [String: String] { get }
     var sort: MyNFTViewModel.Sort? { get set }
     
+    func checkStoredSort()
     func getMyNFTs(nftIDs: [String])
     func toggleLikeFromMyNFT(id: String)
 }
@@ -60,7 +61,6 @@ final class MyNFTViewModel: MyNFTViewModelProtocol {
     }
     
     func getMyNFTs(nftIDs: [String]) {
-        UIBlockingProgressHUD.show()
         var loadedNFTs: [NFTNetworkModel] = []
         
         nftIDs.forEach { id in
@@ -131,7 +131,15 @@ final class MyNFTViewModel: MyNFTViewModelProtocol {
                 return myNFTs.sorted(by: { $0.name < $1.name })
         }
     }
-}
+    
+    func checkStoredSort() {
+            if let sortOrder = UserDefaults.standard.data(forKey: "sortOrder") {
+                let order = try? PropertyListDecoder().decode(MyNFTViewModel.Sort.self, from: sortOrder)
+                sort = order
+            }
+            sort = .rating
+        }
+    }
 
 extension MyNFTViewModel {
     enum Sort: Codable {
